rename module featured-services-intitutions
